# wfzo-api-main/wfzo-api-main/api-spec.yml
openapi: 3.0.4
info:
  title: WFZO Website Management Platform API
  description: >-
    REST APIs for the World Free Zones Organization (WFZO) web platform.
    Scopes
     - Auth: Microsoft Entra (External ID) Bearer JWT
     - Content: read-through to Strapi CMS
     - Documents: secure PDF delivery via short-lived Azure Blob SAS
     - Events: Zoom registration with per-membership seat quotas
     - Learning: LMS integration with Moodle
     - Community: Discourse integrations (DMs, topics, embeds)
     - CMS APIs
  version: 1.0.0
servers:
  - url: /wfzo/api/v1
...
paths:
...
  /member:
    post:
      tags:
        - Member
      summary: Create a new member
      description: Create a new member record; supports basic registration and full onboarding.
      operationId: registration
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MemberInfo"
      responses:
        "200":
          description: successful operation need to respond the memberId
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "400":
          description: Invalid input
        "422":
          description: Validation exception
        default:
          description: Unexpected error
...
  /events:
    post:
      tags:
        - Events
      summary: Create a Zoom-based webinar or meeting
      description: |
        Webhook endpoint called from Strapi CMS when a member or WFZO staff creates an online event. Automatically creates a 
        Zoom webinar/meeting and stores meeting details. Supports per-membership quota settings for registration.
      operationId: createEvent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventCreateRequest"
            examples:
              webinar:
                summary: Create a webinar with per-member quotas
                value:
                  eventCode: "WFZO-2025-DXB"
                  type: "webinar"
                  title: "Innovations in Aerospace"
                  description: "An in-depth look at the latest trends in aerospace technology."
                  scheduledAt: "2025-11-15T14:00:00Z"
                  durationMinutes: 60
                  timezone: "Asia/Dubai"
                  presenter:
                    name: "Dr. Jane Smith"
                    email: "jane@example.com"
                  capacity: 500
                  language: "en"
                  createdBy:
                    memberId: "a4f2f66c-9b0a-4d93-a9e4-3d2f3e6e8a77"
                    userId: "b3e1d44a-8c9b-4f83-b7d3-4e5f6g7h8i9j"
                  registrationSettings:
                    requiresApproval: false
                    registrationDeadline: "2025-11-14T23:59:59Z"
                    allowWaitlist: true
                    perMemberQuota:
                      basic: 2
                      core: 5
                      premium: 10
                  zoom:
                    autoCreate: true
                    type: "webinar"
                    settings:
                      hostVideo: true
                      participantVideo: false
                      muteUponEntry: true
                      waitingRoom: true
      responses:
        "201":
          description: Event created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventCreateResponse"
              example:
                id: "e8f9g0h1-i2j3-k4l5-m6n7-o8p9q0r1s2t3"
                eventCode: "WFZO-2025-DXB"
                status: "scheduled"
                zoom:
                  meetingId: "1234567890"
                  webinarId: "9876543210"
                  hostJoinUrl: "https://zoom.us/s/1234567890?pwd=..."
                  registrationUrl: "https://zoom.us/webinar/register/WN_abc123"
                createdAt: "2025-10-19T10:00:00Z"
        "400":
          description: Invalid input
        "500":
          description: Internal server error
    get:
      tags:
        - Events
      summary: List events
      description: |
        Search and filter events. Supports pagination, status filtering, and text search.
        Returns events with basic details and registration statistics.
      operationId: listEvents
      parameters:
        - name: status
          in: query
          description: Filter by event status
          schema:
            type: string
            enum: [upcoming, ongoing, past, cancelled]
        - name: type
          in: query
          description: Filter by event type
          schema:
            type: string
            enum: [webinar, meeting]
        - name: createdBy
          in: query
          description: Filter by creator member ID
          schema:
            type: string
            format: uuid
        - name: q
          in: query
          description: Text search across title and description
          schema:
            type: string
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/pageSize"
      responses:
        "200":
          description: List of events
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventListResponse"
        "400":
          description: Invalid input
  /events/{eventCode}:
    get:
      tags:
        - Events
      summary: Get event details
      description: |
        Get comprehensive event information including registration status,
        available seats, and user's registration status if authenticated.
      operationId: getEventByCode
      parameters:
        - name: eventCode
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Event details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventDetails"
        "404":
          description: Event not found
    patch:
      tags:
        - Events
      summary: Update event details
      description: |
        Update event information. Restricted to event creator or WFZO admin. Updates are synced to Zoom if applicable.
      operationId: updateEvent
      parameters:
        - name: eventCode
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventUpdateRequest"
      responses:
        "200":
          description: Event updated successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventDetails"
        "400":
          description: Invalid input
        "401":
          description: Unauthorized
        "403":
          description: Forbidden - not event creator or admin
        "404":
          description: Event not found
    delete:
      tags:
        - Events
      summary: Delete event
      description: |
        Delete an event. Only allowed for events with no registrations.
        For events with registrations, use the cancel endpoint instead.
      operationId: deleteEvent
      parameters:
        - name: eventCode
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Event deleted successfully
        "400":
          description: Cannot delete event with registrations
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Event not found
  /events/{eventCode}/cancel:
    post:
      tags:
        - Events
      summary: Cancel an event
      description: |
        Cancel an event and optionally notify all registered attendees.
        Updates Zoom meeting/webinar status and marks all registrations as cancelled.
      operationId: cancelEvent
      parameters:
        - name: eventCode
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [reason]
              properties:
                reason:
                  type: string
                  description: Reason for cancellation
                notifyAttendees:
                  type: boolean
                  default: true
                  description: Send cancellation email to all attendees
      responses:
        "200":
          description: Event cancelled successfully
        "401":
          description: Unauthorized
        "403":
          description: Forbidden
        "404":
          description: Event not found
  /events/{eventCode}/registration/me:
    get:
      tags:
        - Events
      summary: Get my registration status
      description: |
        Get the current user's registration details for a specific event,
        including personalized Zoom join URL.
      operationId: getMyRegistration
      parameters:
        - name: eventCode
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Registration details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MyRegistrationResponse"
        "404":
          description: Not registered or event not found
  /events/registration:
    post:
      tags:
        - Events
      summary: Register for an event
      description: |
        Submit event registration for one or more attendees from the same membership.
        Automatically checks per-member quota limits and creates Zoom registrations.
        Each attendee receives a unique Zoom join URL.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventRegistrationRequest"
            examples:
              singleAttendee:
                summary: Register one attendee
                value:
                  eventCode: "WFZO-2025-DXB"
                  membershipId: "a4f2f66c-9b0a-4d93-a9e4-3d2f3e6e8a77"
                  attendees:
                    - firstName: "Aisha"
                      lastName: "Khan"
                      email: "aisha@example.com"
                      organization: "Alpha Free Zone"
                      jobTitle: "Director"
                      phone: "+971501234567"
                  consent: true
              multipleAttendees:
                summary: Register multiple attendees (batch)
                value:
                  eventCode: "WFZO-2025-DXB"
                  membershipId: "a4f2f66c-9b0a-4d93-a9e4-3d2f3e6e8a77"
                  attendees:
                    - firstName: "Aisha"
                      lastName: "Khan"
                      email: "aisha@example.com"
                      organization: "Alpha Free Zone"
                    - firstName: "Mohammed"
                      lastName: "Ahmed"
                      email: "mohammed@example.com"
                      organization: "Alpha Free Zone"
                  consent: true
      responses:
        "201":
          description: Registration created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventRegistrationResponse"
              example:
                registrationId: "f9cd072d-9b62-4b28-9d72-1b2ad6f93b6b"
                eventCode: "WFZO-2025-DXB"
                status: "confirmed"
                membershipId: "a4f2f66c-9b0a-4d93-a9e4-3d2f3e6e8a77"
                quota:
                  allowed: 5
                  used: 3
                  remaining: 2
                registrations:
                  - attendeeId: "xyz-123"
                    email: "aisha@example.com"
                    status: "confirmed"
                    joinUrl: "https://zoom.us/w/1234?tk=token1"
                    zoomRegistrantId: "abc123"
                  - attendeeId: "xyz-456"
                    email: "mohammed@example.com"
                    status: "confirmed"
                    joinUrl: "https://zoom.us/w/1234?tk=token2"
                    zoomRegistrantId: "def456"
                message: "Successfully registered 2 attendee(s). 2 seats remaining."
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
  /events/registration/details/{eventCode}:
    get:
      tags:
        - Events
      summary: Get event registration details
      description: Retrieve registration details for the given event code (e.g., seat status, registration URL, access info).
      parameters:
        - name: eventCode
          in: path
          description: Unique event code
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Registration details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventRegistrationDetails"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
  /events/registration/{eventCode}/cancel:
    post:
      tags:
        - Events
      summary: Cancel event registration
      description: Cancel the user's registration for the specified event code.
      parameters:
        - name: eventCode
          in: path
          description: Unique event code
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Cancellation successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventRegistrationCancelResponse"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
  /events/registration/{eventCode}/attendees:
    get:
      tags:
        - Events
      summary: List registered attendees for an event
      description: |
        Returns a paginated list of registrations/attendees for the given event code.
        Works for all event types (webinars, member-hosted events, etc.). Supports text search and status filters.
      operationId: getEventAttendees
      parameters:
        - name: eventCode
          in: path
          description: Unique event code
          required: true
          schema:
            type: string
        - name: status
          in: query
          description: Filter by registration status (multi-select)
          required: false
          style: form
          explode: true
          schema:
            type: array
            items:
              $ref: "#/components/schemas/RegistrationStatusFull"
        - name: q
          in: query
          description: Free-text search across attendee name, email, and organization
          required: false
          schema:
            type: string
        - name: includeCancelled
          in: query
          description: Include registrations with status=cancelled in results
          required: false
          schema:
            type: boolean
            default: false
        - $ref: "#/components/parameters/page"
        - $ref: "#/components/parameters/pageSize"
      responses:
        "200":
          description: Paginated attendee list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventAttendeeListResponse"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
  /events/registration/{eventCode}/attendees/export:
    get:
      tags:
        - Events
      summary: Export attendees as Excel
      description: |
        Downloads the attendees list as an Excel file (.xlsx). Applies same filters as the list endpoint.
      operationId: exportEventAttendees
      parameters:
        - name: eventCode
          in: path
          description: Unique event code
          required: true
          schema:
            type: string
        - name: status
          in: query
          description: Filter by registration status (multi-select)
          required: false
          style: form
          explode: true
          schema:
            type: array
            items:
              $ref: "#/components/schemas/RegistrationStatusFull"
        - name: q
          in: query
          description: Free-text search across attendee name, email, and organization
          required: false
          schema:
            type: string
        - name: includeCancelled
          in: query
          description: Include registrations with status=cancelled in results
          required: false
          schema:
            type: boolean
            default: false
      responses:
        "200":
          description: Excel export
          headers:
            Content-Disposition:
              schema:
                type: string
              description: Attachment disposition including filename (e.g., attachment; filename="attendees.xlsx")
          content:
            application/vnd.openxmlformats-officedocument.spreadsheetml.sheet:
              schema:
                type: string
                format: binary
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
  /events/send-email:
    post:
      tags:
        - Events
      summary: Send event notification email
      description: |
        Send a templated email for event approval workflow notifications.
        Supports 4 different email types for event submission and approval processes.
        EVENT_SUBMITTED_FOR_APPROVAL sends two emails: one to admin and one to user.
      operationId: sendEventEmail
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SendEventEmailRequest"
            examples:
            eventSubmitted:
              summary: Send both admin notification and user confirmation when event is submitted
              value:
                email: "organizer@example.com"
                type: "EVENT_SUBMITTED_FOR_APPROVAL"
                eventCode: "WFZO-2025-DXB"
                eventTitle: "Innovations in Aerospace"
                presenterName: "Dr. Jane Smith"
                scheduledDate: "2025-11-15T14:00:00Z"
                eventType: "webinar"
                capacity: 500
                firstName: "John"
                lastName: "Doe"
              eventApproved:
                summary: Notify user when event is approved
                value:
                  email: "organizer@example.com"
                  type: "EVENT_APPROVED_USER"
                  eventCode: "WFZO-2025-DXB"
                  eventTitle: "Innovations in Aerospace"
                  scheduledDate: "2025-11-15T14:00:00Z"
                  eventType: "webinar"
                  capacity: 500
                  firstName: "John"
                  lastName: "Doe"
              eventRejected:
                summary: Notify user when event is rejected
                value:
                  email: "organizer@example.com"
                  type: "EVENT_REJECTED_USER"
                  eventCode: "WFZO-2025-DXB"
                  eventTitle: "Innovations in Aerospace"
                  scheduledDate: "2025-11-15T14:00:00Z"
                  rejectionReason: "Event conflicts with existing schedule"
                  firstName: "John"
                  lastName: "Doe"
      responses:
        "200":
          description: Email sent successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SendEventEmailResponse"
              example:
                message: "Email sent successfully"
        "400":
          description: Invalid input
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiResponse"
        "500":
          description: Internal server error
...
components:
  parameters:
    page:
      in: query
      name: page
      description: Page number for pagination (1-based)
      schema:
        type: integer
        minimum: 1
        default: 1
...
  schemas:
...
    MemberInfo:
      type: object
      description: Membership attributes and status for a member organization
      required: [category, status]
      properties:
        id:
          type: string
          format: uuid
        users:
          type: array
          items: { $ref: "#/components/schemas/User" }
        Address:
          $ref: "#/components/schemas/Address"
        category:
          type: string
          $ref: "#/components/schemas/MembershipCategory"
          description:
            Membership category. This is the basic Membership level. Current system supports 3
            categories votingMember, associateMember, partnerObserver with no additional tier. This is expected
            to be extended to 6 with multiple tier. We will also control the categories to be displayed or available
            in the website using a flag(active/inactive).
        tier:
          type: string
          $ref: "#/components/schemas/Tier"
          description:
            Membership tier within the category. This is expected to be extended to 3 tiers - basic, core, premium.
            We will also control the tiers to be displayed or available in the
        status:
          type: string
          enum: [active, expired, pendingApproval, pendingFormSubmission]
        validUntil:
          type: string
          format: date-time
          nullable: true
        OrganisationInfo:
          $ref: "#/components/schemas/OrganisationInfo"
        MemberConsent:
          $ref: "#/components/schemas/MemberConsent"
        featuredMember:
          type: boolean
          description: Whether to feature this member on the website
...
    EventCreateResponse:
      type: object
      description: Response after creating an event
      required: [id, eventCode, status, createdAt]
      properties:
        id: { type: string, format: uuid }
        eventCode: { type: string }
        status:
          $ref: "#/components/schemas/EventStatus"
        zoom:
          $ref: "#/components/schemas/ZoomDetails"
        createdAt: { type: string, format: date-time }
...
    ApiResponse:
      type: object
      description: Standard API response envelope for status reporting
      properties:
        code:
          type: integer
          format: int32
        type:
          type: string
        message:
          type: string
      xml:
        name: "##default"
...
  securitySchemes:
    petstore_auth:
      type: oauth2
      flows:
        implicit:
          authorizationUrl: https://petstore3.swagger.io/oauth/authorize
          scopes:
            write:pets: modify pets in your account
            read:pets: read your pets
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    api_key:
      type: apiKey
      name: api_key
      in: header